requirements-dev.txt
ansicon==1.89.0; platform_system == "Windows"
atomicwrites==1.4.0; python_version >= "3.6" and python_full_version < "3.0.0" and sys_platform == "win32" or sys_platform == "win32" and python_version >= "3.6" and python_full_version >= "3.4.0"
attrs==20.3.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
awscli==1.20.13; python_version >= "3.6"
blessed==1.17.6
boto3-stubs==1.18.13; python_version >= "3.6"
boto3==1.18.13; python_version >= "3.6"
botocore-stubs==1.21.13; python_version >= "3.6"
botocore==1.21.13; python_version >= "3.6"
certifi==2021.5.30; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
cffi==1.14.6; python_version >= "3.6"
chalice==1.24.1
charset-normalizer==2.0.4; python_full_version >= "3.6.0" and python_version >= "3"
click==7.1.2; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
colorama==0.4.3; python_version >= "3.6" and python_full_version < "3.0.0" and sys_platform == "win32" or python_full_version >= "3.5.0" and python_version >= "3.6" and sys_platform == "win32"
cryptography==3.4.7; python_version >= "3.6"
dataclasses==0.8; python_version >= "3.6" and python_version < "3.7"
docutils==0.15.2; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
dsnap==1.0.0; python_version >= "3.6" and python_version < "4.0"
flake8==3.9.2; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
freezegun==1.1.0; python_version >= "3.5"
idna==3.2; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.5"
importlib-metadata==4.6.3; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.8" or python_full_version >= "3.5.0" and python_version < "3.8" and python_version >= "3.6"
iniconfig==1.1.1; python_version >= "3.6"
inquirer==2.7.0
jinja2==3.0.1; python_version >= "3.6"
jinxed==1.1.0; platform_system == "Windows"
jmespath==0.10.0; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "4.0" or python_full_version >= "3.3.0" and python_version >= "3.6" and python_version < "4.0"
markupsafe==2.0.1; python_version >= "3.6"
mccabe==0.6.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
more-itertools==8.8.0; python_version >= "3.5"
moto==2.2.1
mypy-boto3-iam==1.18.13; python_version >= "3.6"
mypy-boto3-lambda==1.18.13; python_version >= "3.6"
mypy-boto3-s3==1.18.13; python_version >= "3.6"
mypy-extensions==0.4.3; python_version >= "3.5"
mypy==0.812; python_version >= "3.5"
packaging==21.0; python_version >= "3.6"
pluggy==0.13.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
py==1.10.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
pyasn1==0.4.8; python_version >= "3.6" and python_version < "4"
pycodestyle==2.7.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
pycparser==2.20; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
pyflakes==2.3.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
pyparsing==2.4.7; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
pytest==6.2.4; python_version >= "3.6"
python-dateutil==2.8.2; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
python-editor==1.0.4
pytz==2021.1
pyyaml==5.4.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
readchar==2.0.1
requests==2.26.0; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
responses==0.13.3; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
rsa==4.7.2; python_version >= "3.6" and python_version < "4"
s3transfer==0.5.0; python_version >= "3.6" and python_version < "4.0"
six==1.16.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
sqlalchemy-stubs==0.4
sqlalchemy-utils==0.36.8
sqlalchemy==1.3.24; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0")
toml==0.10.2; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
typed-ast==1.4.3; python_version >= "3.5"
typing-extensions==3.10.0.0
typing==3.6.4; python_version < "3.7"
urllib3==1.26.6; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0" and python_version < "4")
wcwidth==0.2.5
werkzeug==2.0.1; python_version >= "3.6"
xmltodict==0.12.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
zipp==3.5.0; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.8" or python_full_version >= "3.5.0" and python_version < "3.8" and python_version >= "3.6"
